Settings: Add .exclude_from, paths containing exclude patterns
This exposes the DirectoryTraversal.exclude_from() API to the Settings
layer in FawltyDeps, and allows configuration of which files we read
exclude patterns from.

When not customized, this new setting defaults to reading ignore
patterns from these files:

 - .ignore
 - .gitignore
 - .git/info/exclude

This default is inspired by Ruff, which has a `respect-gitignore`
directive that defaults to `true`, and that observes these files.
(See https://docs.astral.sh/ruff/settings/#respect-gitignore for more
details.)

As with other Settings directives, customizing the exclude_from member
will _replace_ (not extend) the defaults.