real_projects: Rewrite python-algorithms test to work on Windows
The python-algortihms project highlights an outstanding issue in
FawltyDeps: Placeholder packages. As discussed in issue #176, these are
packages that don't provide the expected import name (or any import name
at all), but instead depend on other packages (transitive dependencies)
which will then supply the relevant import name.

The first example we came across of such a package (in #176) was qiskit.
Another example is tensorflow, but only on Windows: On Windows the
tensorflow package is a 1.9kB placeholder that depends on
tensorflow-intel (266MB) to bring in the actual meat. (On POSIX, the
tensorflow package itself contains everything.)

We make two changes to our real_projects test for python-algorithms:

1. We add the posix_only flag to the all_reqs_installed experiment, to
   prevent it from running on Windows. Its expected results would be
   different there (because of tensorflow-intel).

2. We add a new experiment, some_reqs_customized, which uses a
   custom_mapping to resolve qiskit and tensorflow. The remaining
   requirements are installed (as in the previous experiment).
   This yields the "best" FawltyDeps result for this project.
   The remaining undeclared/unused dependencies are presumed to be
   true positives for this project.