test_sample_projects: Run experiments from within the project dir
Instead of passing absolute paths to the project dir, chdir into the
project_dir and run the experiment with relative paths instead.

This is closer to how users run FawltyDeps in their own projects, and is
also necessary in order to properly test --exclude-from's default
behavior (which will look for .ignore, .gitignore, and .git/info/exclude
under the _current_ directory).